Add set tracee registers
The tracer was used to set the tracee's rip register which allowed any
code within the tracee to be executed. The address is computed based on
the base map address for the text section from
`/proc/{tracee_pid}/maps` and the instruction offset we want to set the
rip at.